on_data: Always parse the whole view
With the current approach of only parsing new data there could be data
races such as when the size of the view isn't updated in time which
would result in some regions being at the wrong position.

This fixes the above problem by running the colorizing command on the
whole view on every update.